split monolithic meteor.py into multiple files
